Refactor examples to use shared utils for port selection and improve async stream handling
