+++ examples/ping/ping.py
@@ -1,10 +1,11 @@
 import argparse
-import random
-import socket
 
 import multiaddr
 import trio
 
+from examples.utils import (
+    get_random_available_port,
+)
 from libp2p import (
     new_host,
 )
@@ -98,7 +99,6 @@ def main() -> None:
     )
 
     parser = argparse.ArgumentParser(description=description)
-
     parser.add_argument("-p", "--port", type=int, help="source port number (optional)")
     parser.add_argument(
         "-d",
@@ -113,23 +113,6 @@ def main() -> None:
         trio.run(run, *(port, args.destination))
     except KeyboardInterrupt:
         pass
-
-
-def is_port_available(port: int) -> bool:
-    with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as sock:
-        try:
-            sock.bind(("", port))
-            return True
-        except OSError:
-            return False
-
-
-def get_random_available_port(start: int = 10000, end: int = 20000) -> int:
-    for _ in range(50):  # try up to 50 random ports
-        port = random.randint(start, end)
-        if is_port_available(port):
-            return port
-    raise RuntimeError("Could not find an available port in the given range.")
 
 
 if __name__ == "__main__":
